fix: use lecture id to start liveroom
